[refactor] : 일관된 순서로 routes, App.js 경로 수정 & mypage 폴더에 관련 페이지 카테고리화
