+++ src/constant/routes.js
@@ -3,18 +3,18 @@ const routes = {
   login: '/login',
   loginKakao: '/login/callback',
   registerBank: '/register/bank',
-  mypage: '/mypage',
-  writenPost: '/mypage/writen-post',
-  pickupPost: '/mypage/pickup-post',
-  writenPostDetail: '/mypage/writen-post/:id',
-  pickupPostDetail: '/mypage/pickup-post/:id',
-  admin: '/mypage/admin',
-  adminAuth: '/mypage/admin/auth/:id',
-  uploadStudentCard: '/mypage/uploadStudentCard',
   post: '/post',
   detailPost: '/post/:id',
   postWriteIntro: '/post-write-intro',
   postWrite: '/post-write',
+  mypage: '/mypage',
+  admin: '/mypage/admin',
+  adminAuth: '/mypage/admin/auth/:id',
+  writenPost: '/mypage/writen-post',
+  pickupPost: '/mypage/pickup-post',
+  writenPostDetail: '/mypage/writen-post/:id',
+  pickupPostDetail: '/mypage/pickup-post/:id',
+  uploadStudentCard: '/mypage/uploadStudentCard',
   error: '/*',
 };
