+++ src/pages/mypage/MyPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MyPageTemplate from '../components/templates/MyPageTemplate';
+import MyPageTemplate from '../../components/templates/MyPageTemplate';
 
 const MyPage = () => {
   return <MyPageTemplate />;
